move hmac/cmac to mac module, move MACContext to mac module
