+++ docs/hazmat/primitives/mac/index.rst
@@ -16,3 +16,4 @@ HMAC?`_
 
     cmac
     hmac
+    interfaces
