+++ src/cryptography/fernet.py
@@ -16,7 +16,7 @@ from cryptography.exceptions import InvalidSignature
 from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives import hashes, padding
 from cryptography.hazmat.primitives.ciphers import Cipher, algorithms, modes
-from cryptography.hazmat.primitives.hmac import HMAC
+from cryptography.hazmat.primitives.mac.hmac import HMAC
 
 
 class InvalidToken(Exception):
